My first commit.
The HandlerKeys class generate a pair of keys(public and private).
Using rsa as cryptographic algorithm but it is can be easily changed.
The length and location of keys are storage in configuration file.
The generated keys are stored in public.key and private.key files.

Willian, I love You too. <3